requirements.txt
fastapi
huggingface_hub
langdetect
Pillow
pydantic
Requests
torch
torchvision
transformers